apps/lambda_py/requirements-dev.txt
-r requirements.txt
pytest>=8.3.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
moto>=5.0.0
black>=24.10.0
ruff>=0.8.0
mypy>=1.13.0
boto3-stubs[lambda,dynamodb,s3]>=1.35.0
